tests/InFlux_test.py
import unittest
from unittest.mock import patch, MagicMock
from ..app.connectors.InFlux import InFlux  # replace 'your_module' with the name of your module

class TestInFlux(unittest.TestCase):

    def setUp(self):
        self.influx = InFlux()

    @patch('your_module.influxdb_client.InfluxDBClient')
    def test_connect_(self, mock_client):
        # Test successful connection
        self.assertTrue(self.influx.connect_("mtu", "localhost", "8086", "6R9Hg8swLRMBSVn6swE6yqmXAhSdcHAZ_G73NjP6QhdQKUkYhZWEuRfP-WsDKg5A3aYSnpBzmcG4fkUEoxZeGQ=="))
        # You may also want to check if InfluxDBClient was called with the right arguments
        
        # Test connection failure
        mock_client.side_effect = Exception("Connection Failed")
        self.assertFalse(self.influx.connect_("org", "host", "port", "token"))


    def test_connect(self):

        mapping={}
        connection_profile={"org":"mtu", "host": "localhost", "port":"8086", "token":"6R9Hg8swLRMBSVn6swE6yqmXAhSdcHAZ_G73NjP6QhdQKUkYhZWEuRfP"}

    
        self.assertTrue(self.influx.connect(mapping=mapping, connection_profile=connection_profile))

    # def test_disconnect(self):
    #     # Test if disconnect calls the right method in the InfluxDB client
    #     pass

    # @patch('your_module.InFlux.connection')
    # def test_get_latest_data(self, mock_connection):
    #     # Mock the cursor method and its return values to test get_latest_data
    #     pass

    # @patch('your_module.InFlux.connection')
    # def test_get_data(self, mock_connection):
    #     # Mock the cursor method and its return values to test get_data (SQL version)
    #     pass

    # @patch('your_module.InFlux.connection')
    # def test_get_data_influx(self, mock_connection):
    #     # Mock the query_api and its return values to test get_data (InfluxDB version)
    #     pass

    # def test_post_data(self):
    #     # Depending on what post_data is supposed to do, write tests for it
    #     pass

    # @patch('your_module.InFlux.connection')
    # def test_describe_db(self, mock_connection):
    #     # Mock the cursor and its methods to test describe_db
    #     pass

if __name__ == '__main__':
    unittest.main()